% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSeqs.R
\name{exportBins}
\alias{exportBins}
\title{Export the bins of a SQM object}
\usage{
exportBins(SQM, output_dir = "")
}
\arguments{
\item{SQM}{A SQM object.}

\item{output_dir}{Existing output directory to which to write the bins.}
}
\description{
Export the bins of a SQM object
}
