% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSeqs.R
\name{exportContigs}
\alias{exportContigs}
\title{Export the contigs of a SQM object}
\usage{
exportContigs(SQM, output_name = "")
}
\arguments{
\item{SQM}{A SQM object.}

\item{output_name.}{A connection, or a character string naming the file to print to. If "" (the default), sequences will be printed to the standard output connection.}
}
\description{
Export the contigs of a SQM object
}
