% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportPathway.R
\name{exportPathway}
\alias{exportPathway}
\title{Export the functions of a SQM object into KEGG pathway maps}
\usage{
exportPathway(
  SQM,
  pathway_id,
  samples = NULL,
  split_samples = F,
  output_suffix = "pathview"
)
}
\arguments{
\item{SQM}{A SQM or SQMlite object.}

\item{pathway_id}{character. The five-number KEGG pathway identifier. A list of all pathway identifiers can be found in \url{https://www.genome.jp/kegg/pathway.html}.}

\item{samples}{character. A vector with the names of the samples to export. If absent, all samples will be exported (default \code{NULL}).}

\item{split_samples}{logical. Generate a different output file for each sample (default \code{FALSE}).}

\item{output_suffix}{character. Suffix to be added to the output files (default \code{"pathview"}).}
}
\description{
This function is a wrapper for the pathview package (Luo \emph{et al.}, 2017. \emph{Nucleic acids research}, 45:W501-W508). It will generate annotated KEGG pathway maps showing which reactions are present in the different samples.
}
\examples{
data(Hadza)
exportPathway(Hadza, "00910", output_suffix = "nitrogen_metabolism")
}
\seealso{
\code{\link[plotFunctions]{plotFunctions}} for plotting the most functions taxa of a SQM object.
}
