% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTable.R
\name{exportTable}
\alias{exportTable}
\title{Export results in tabular format}
\usage{
exportTable(table, output_name)
}
\arguments{
\item{table}{vector, matrix or data.frame. The table to be written.logical.}

\item{output_name}{character. Name of the output file.}
}
\description{
This function is a wrapper for R's write.table function.
}
\examples{
data(Hadza)
Hadza.iron = subsetFun(Hadza, "iron")
# Write the taxonomic distribution at the genus level of all the genes related to iron.
exportTable(Hadza.iron$taxa$genus$percent, "Hadza.ironGenes.genus.tsv")
# Now write the distribution of the different iron-related COGs (Clusters of Orthologous Groups) across samples.
exportTable(Hadza.iron$functions$COG$tpm, "Hadza.ironGenes.COG.tsv")
# Now write all the information contained in the ORF table.
exportTable(Hadza.iron$orfs$table, "Hadza.ironGenes.orftable.tsv")
}
