% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecA.R
\docType{data}
\name{RecA}
\alias{RecA}
\title{RecA/RadA recombinase}
\format{Character vector with the COG identifier for RecA/RadA.}
\source{
\href{http://eggnogdb.embl.de/#/app/results?seqid=P0A7G6&target_nogs=COG0468#COG0468_datamenu}{EggNOG Database}.
}
\usage{
data(RecA)
}
\description{
The recombination protein RecA/RadA is essential for the repair and
maintenance of DNA, and has homologs in every bacteria and archaea.
By dividing the coverage of functions by the coverage of RecA, abundances
can be transformed into copy numbers, which can be used to compare 
functional profiles in samples with different sequencing depths.
RecA-derived copy numbers are available in the SQM object
(\code{SQM$functions$<annotation_type>$copy_number}).
}
\examples{
data(Hadza)
data(RecA)
### Let's calculate the average copy number of each function in our samples.
# We do it for COG annotations here, but we could also do it for KEGG or PFAMs.
COG.coverage = SQMtools::aggregate.fun(Hadza, 'COG', trusted_functions_only=T, ignore_unclassified_functions=F)$cov
COG.copynumber = t(t(COG.coverage) / COG.coverage[RecA,]) # Sample-wise division by RecA tpm.
}
\keyword{datasets}
