% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_methods.R
\name{remove_contigs_from_bin}
\alias{remove_contigs_from_bin}
\title{Remove contigs from a given bin}
\usage{
remove_contigs_from_bin(SQM, bin, contigs)
}
\arguments{
\item{SQM}{A SQM object.}

\item{bin}{character. Name of the bin from which the contigs will be removed.}

\item{contigs}{character. Vector with the names of the contigs that will be removed from the new bin.}
}
\description{
Remove contigs from a given bin
}
\seealso{
\code{\link{find_redundant_contigs}}, \code{\link{create_bin}}
}
