% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_methods.R
\name{renameSamples}
\alias{renameSamples}
\title{Change sample names}
\usage{
renameSamples(SQM, new_sample_names)
}
\arguments{
\item{SQM}{SQM or SQMlite object}

\item{new_sample_names}{character. New sample names}
}
\value{
SQM or SQMlite object with the new sample names
}
\description{
Change the sample names of a SQM or SQMlite object
}
