% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_methods.R
\name{rowMaxs}
\alias{rowMaxs}
\title{Return a vector with the row-wise maxima of a matrix or dataframe.}
\usage{
rowMaxs(table)
}
\arguments{
\item{table}{matrix or dataframe.}
}
\value{
a vector with the row-wise maxima.
}
\description{
Return a vector with the row-wise maxima of a matrix or dataframe.
}
