% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_methods.R
\name{rowMins}
\alias{rowMins}
\title{Return a vector with the row-wise minima of a matrix or dataframe.}
\usage{
rowMins(table)
}
\arguments{
\item{table}{matrix or dataframe.}
}
\value{
a vector with the row-wise minima.
}
\description{
Return a vector with the row-wise minima of a matrix or dataframe.
}
