% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_methods.R
\name{subsetBins}
\alias{subsetBins}
\title{Create a SQM object containing only the requested bins, and the contigs and ORFs contained in them.}
\usage{
subsetBins(
  SQM,
  bins = NULL,
  rank = NULL,
  tax = NULL,
  min_completeness = NULL,
  max_contamination = NULL,
  tax_source = "bins",
  trusted_functions_only = FALSE,
  ignore_unclassified_functions = FALSE,
  rescale_tpm = TRUE,
  rescale_copy_number = TRUE,
  allow_empty = FALSE
)
}
\arguments{
\item{SQM}{SQM object to be subsetted.}

\item{bins}{character. Vector of bins to be selected. If provided, will override \code{rank}, \code{tax}, \code{min_completeness} and \code{max_contamination}.}

\item{rank}{character. The taxonomic rank from which to select the desired taxa (\code{superkingdom}, \code{phylum}, \code{class}, \code{order}, \code{family}, \code{genus}, \code{species})}

\item{tax}{character. A taxon or vector of taxa to be selected.}

\item{min_completeness}{numeric. Discard bins with completeness lower than this value (default \code{NULL}).}

\item{max_contamination}{numeric. Discard bins with contamination higher than this value (default \code{NULL}).}

\item{tax_source}{character, source data used for taxonomic subsetting (if \code{rank} and \code{tax} are provided) and for the aggregate taxonomy tables present in \code{SQM$taxa}, either \code{"orfs"}, \code{"contigs"}, \code{"bins"} (GTDB bin taxonomy if available, SQM bin taxonomy otherwise), \code{"bins_gtdb"} (GTDB bin taxonomy) or \code{"bins_sqm"} (SQM bin taxonomy). If using \code{bins_gtdb}, note that GTDB taxonomy may differ from the NCBI taxonomy used throughout the rest of SqueezeMeta. Default \code{"bins"}.}

\item{trusted_functions_only}{logical. If \code{TRUE}, only highly trusted functional annotations (best hit + best average) will be considered when generating aggregated function tables. If \code{FALSE}, best hit annotations will be used (default \code{FALSE}).}

\item{ignore_unclassified_functions}{logical. If \code{FALSE}, ORFs with no functional classification will be aggregated together into an "Unclassified" category. If \code{TRUE}, they will be ignored (default \code{FALSE}).}

\item{rescale_tpm}{logical. If \code{TRUE}, TPMs for KEGGs, COGs, and PFAMs will be recalculated (so that the TPMs in the subset actually add up to 1 million). Otherwise, per-function TPMs will be calculated by aggregating the TPMs of the ORFs annotated with that function, and will thus keep the scaling present in the parent object. By default it is set to \code{TRUE}, which means that the returned TPMs will be scaled \emph{by million of reads of the selected bins}.}

\item{rescale_copy_number}{logical. If \code{TRUE}, copy numbers with be recalculated using the median single-copy gene coverages in the subset. Otherwise, single-copy gene coverages will be taken from the parent object. By default it is set to \code{TRUE}, which means that the returned copy numbers for each function will represent the average copy number of that function \emph{per genome of the selected taxon}.}

\item{allow_empty}{(internal use only).}
}
\value{
SQM object containing only the requested bins.
}
\description{
Create a SQM object containing only the requested bins, and the contigs and ORFs contained in them.
}
\examples{
data(Hadza)
# Which are the most complete bins?
topBinNames = rownames(Hadza$bins$table)[order(Hadza$bins$table[,"Completeness"],
                                         decreasing=TRUE)][1:2]
# Subset with the most complete bin.
topBin = subsetBins(Hadza, topBinNames[1])

# Subset with all the bins over 90\% completeness
over90 = subsetBins(Hadza, min_completeness = 90)

# Subset with bins from the Phascolarctobacterium genus using SqueezeMeta's taxonomy
phasco = subsetBins(Hadza, tax_source = "bins_sqm", rank = "genus", tax = "Phascolarctobacterium")

# Subset with binsfrom the Bacteroidota phylum using GTDB taxonomy
bact = subsetBins(Hadza, tax_source = "bins_gtdb", rank = "phylum", tax = "Bacteroidota")

}
\seealso{
\code{\link{subsetContigs}}, \code{\link{subsetORFs}}
}
