% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_methods.R
\name{subsetRand}
\alias{subsetRand}
\title{Select random ORFs}
\usage{
subsetRand(SQM, N)
}
\arguments{
\item{SQM}{SQM object to be subsetted.}

\item{N}{numeric. number of random ORFs to select.}
}
\value{
SQM object containing a random subset of ORFs.
}
\description{
Create a random subset of a SQM object.
}
\seealso{
\code{\link{subsetORFs}}
}
