% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_methods.R
\name{subsetTax}
\alias{subsetTax}
\title{Filter results by taxonomy}
\usage{
subsetTax(SQM, rank, tax, trusted_functions_only = F,
  ignore_unclassified_functions = F, rescale_copy_number = T)
}
\arguments{
\item{SQM}{SQM object to be subsetted.}

\item{rank}{character. The taxonomic rank from which to select the desired taxa (\code{superkingdom}, \code{phylum}, \code{class}, \code{order}, \code{family}, \code{genus}, \code{species})}

\item{tax}{character. The taxon to select.}

\item{trusted_functions_only}{logical. If \code{TRUE}, only highly trusted functional annotations (best hit + best average) will be considered when generating aggregated function tables. If \code{FALSE}, best hit annotations will be used (default \code{FALSE}).}

\item{ignore_unclassified_functions}{logical. If \code{FALSE}, ORFs with no functional classification will be aggregated together into an "Unclassified" category. If \code{TRUE}, they will be ignored (default \code{FALSE}).}

\item{rescale_copy_number}{logical. If \code{TRUE}, copy numbers with be recalculated using the RecA/RadA coverages in the subset. Otherwise, RecA/RadA coverages will be taken from the original object. By default it is set to \code{TRUE}, which means that the returned copy numbers will represent the average copy number per function \emph{in the genomes of the selected taxa}.}
}
\value{
SQM object containing only the requested taxon.
}
\description{
Create a SQM object containing only the contigs with a given consensus taxonomy, the ORFs contained in them and the bins that contain them.
}
\examples{
data(Hadza)
Hadza.Escherichia = subsetTax(Hadza, 'genus', 'Escherichia')
Hadza.Bacteroidetes = subsetTax(Hadza, 'phylum', 'Bacteroidetes')
}
\seealso{
\code{\link[subsetFun]{subsetFun}}, \code{\link[subsetContigs]{subsetContigs}}, \code{\link[combineSQM]{combineSQM}}. The most abundant items of a particular table contained in a SQM object can be eselected with \code{\link[mostAbundant]{mostAbundant}}.
}
