% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.SQM}
\alias{summary.SQM}
\title{summary method for class SQM}
\usage{
\method{summary}{SQM}(SQM)
}
\value{
A list of summary statistics.
}
\description{
Computes different statistics of the data contained in the SQM object.
}
