% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bunch.R
\name{summary.SQMbunch}
\alias{summary.SQMbunch}
\title{summary method for class SQMbunch}
\usage{
\method{summary}{SQMbunch}(object, ...)
}
\arguments{
\item{object}{SQMbunch object to be summarized.}

\item{...}{Additional parameters (ignored).}
}
\value{
A list of summary statistics.
}
\description{
Computes different statistics of the data contained in the SQMbunch object.
}
