% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_lite.R
\name{summary.SQMlite}
\alias{summary.SQMlite}
\title{summary method for class SQMlite}
\usage{
\method{summary}{SQMlite}(object, ...)
}
\arguments{
\item{object}{SQMlite object to be summarized.}

\item{...}{Additional parameters (ignored).}
}
\value{
A list of summary statistics.
}
\description{
Computes different statistics of the data contained in the SQMlite object.
}
