% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineSQM.R
\name{combineSQM}
\alias{combineSQM}
\title{Combine several SQM objects}
\usage{
combineSQM(..., tax_source = "orfs", trusted_functions_only = F,
  ignore_unclassified_functions = F, rescale_copy_number = T)
}
\arguments{
\item{...}{an arbitrary number of SQM objects}

\item{tax_source}{character. Features used for calculating aggregated abundances at the different taxonomic ranks. Either \code{"orfs"} or \code{"contigs"} (default \code{"orfs"}). If the objects being combined contain a subset of taxa or bins, this parameter can be set to \code{TRUE}.}

\item{trusted_functions_only}{logical. If \code{TRUE}, only highly trusted functional annotations (best hit + best average) will be considered when generating aggregated function tables. If \code{FALSE} (default), best hit annotations will be used.}

\item{ignore_unclassified_functions}{logical. If \code{FALSE}, ORFs with no functional classification will be aggregated together into an "Unclassified" category. If \code{TRUE}, they will be ignored.}

\item{rescale_copy_number}{logical. If \code{TRUE}, copy numbers with be recalculated using the RecA/RadA coverages in the subset. Otherwise, RecA/RadA coverages will be taken from the original object. By default it is set to \code{TRUE}, which means that the returned copy numbers will represent the average copy number per function \emph{in the genomes of the selected bins or contigs}. If any SQM objects that are being combined contain a functional subset rather than a contig/bins subset, this parameter should be set to \code{FALSE}.}
}
\value{
A SQM object
}
\description{
Combine an arbitrary number of SQM objects into a single SQM object.
}
\examples{
data(Hadza)
# Select Carbohydrate metabolism ORFs in Bacteroidetes, and Amino acid metabolism ORFs in Proteobacteria
bact = subsetTax(Hadza, 'phylum', 'Bacteroidetes <phylum>')
bact.carb = subsetFun(bact, 'Carbohydrate metabolism')
proteo = subsetTax(Hadza, 'phylum', 'Proteobacteria')
proteo.amins = subsetFun(proteo, 'Amino acid metabolism')
bact.carb_proteo.amins = combineSQM(bact.carb, proteo.amins, rescale_copy_number=F)
}
\seealso{
\code{\link[subsetFun]{subsetFun}}, \code{\link[subsetTax]{subsetTax}}
}
