% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqvec2fasta.R
\name{seqvec2fasta}
\alias{seqvec2fasta}
\title{Print a named vector of sequences as a fasta-formatted string}
\usage{
seqvec2fasta(seqvec)
}
\arguments{
\item{seqvec}{vector. The vector to be written as a fasta string.}
}
\value{
a character string with the sequence/s written in the fasta format.
}
\description{
Print a named vector of sequences as a fasta-formatted string
}
\examples{
data(Hadza)
seqvec2fasta(Hadza$orfs$seqs[1:10])
}
