% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_methods.R
\name{subsetBins}
\alias{subsetBins}
\title{Create a SQM object containing only the requested bins, and the contigs and ORFs contained in them.}
\usage{
subsetBins(SQM, bins, trusted_functions_only = F,
  ignore_unclassified_functions = F, rescale_copy_number = T)
}
\arguments{
\item{SQM}{SQM object to be subsetted.}

\item{bins}{character. Vector of bins to be selected.}

\item{trusted_functions_only}{logical. If \code{TRUE}, only highly trusted functional annotations (best hit + best average) will be considered when generating aggregated function tables. If \code{FALSE}, best hit annotations will be used (default \code{FALSE}).}

\item{ignore_unclassified_functions}{logical. If \code{FALSE}, ORFs with no functional classification will be aggregated together into an "Unclassified" category. If \code{TRUE}, they will be ignored (default \code{FALSE}).}

\item{rescale_copy_number}{logical. If \code{TRUE}, copy numbers with be recalculated using the RecA/RadA coverages in the subset. Otherwise, RecA/RadA coverages will be taken from the original object. By default it is set to \code{TRUE}, which means that the returned copy numbers will represent the average copy number per function \emph{in the genomes of the selected bins}.}
}
\value{
SQM object containing only the requested bins.
}
\description{
Create a SQM object containing only the requested bins, and the contigs and ORFs contained in them.
}
\examples{
data(Hadza)
# Which are the five most complete bin?
topBinNames = rownames(Hadza$bins$table)[order(Hadza$bins$table[,'Completeness'], decreasing=T)][1:5]
topBins = subsetBins(Hadza, topBinNames)
}
\seealso{
\code{\link[subsetContigs]{subsetContigs}}, \code{\link[subsetORFs]{subsetORFs}}
}
