% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_methods.R
\name{subsetSamples}
\alias{subsetSamples}
\title{Filter results by sample}
\usage{
subsetSamples(SQM, samples, remove_missing = T)
}
\arguments{
\item{SQM}{SQM object to be subsetted.}

\item{samples}{character. Samples to be included in the subset.}

\item{remove_missing}{bool. If \code{TRUE}, ORFs, contigs, bins, taxa and functions absent from the selected samples will be removed from the subsetted object (default \code{TRUE}).}
}
\value{
SQM object containing only the requested samples.
}
\description{
Create a SQM object containing only samples specified by the user, and the ORFs, contigs, bins, taxa and functions present in those samples.
}
\seealso{
\code{\link[subsetTax]{subsetTax}}, \code{\link[subsetFun]{subsetFun}}, \code{\link[subsetORFs]{subsetORFs}}, \code{\link[combineSQM]{combineSQM}}. The most abundant items of a particular table contained in a SQM object can be selected with \code{\link[mostAbundant]{mostAbundant}}.
}
